requirements.txt
aniso8601==9.0.1
autopep8==2.0.2
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-JWT-Extended==4.4.4
Flask-RESTful==0.3.9
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
protobuf==3.20.3
pycodestyle==2.10.0
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2023.2
six==1.16.0
Werkzeug==2.2.3